pyproject.toml
[project]
name = "maieo"
version = "0.1.0"
description = "Maieo AI — Multi-Agent Adversarial Reasoning Engine"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "openai",
    "google-genai",
    "python-dotenv",
    "rich",
    "pydantic",
    "groq",
    "requests",
]
requires-python = ">=3.9"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project.scripts]
maieo = "maieo.cli:main"

